+++ CHANGES.md
@@ -2,7 +2,8 @@
 
 ## 4.3.2
 
-* Correct mem free issue when writing CRAM in mismatchQc
+* Add options for input/output of CRAM in mismatchQc and mmFlagModifier
+* Correct mem free issue when writing CRAM in mismatchQc and mmFlagModifier
 * Resolved #40 - faster recovery times
 * Resolved #39 - handle splitting correctly